% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripe.R
\name{lazy}
\alias{lazy}
\title{Convert magrittr chain lazy function}
\usage{
lazy(x)
}
\arguments{
\item{x}{magrittr chain of functions}
}
\value{
function
}
\description{
Input a chain in magrittr syntax and make it a lazyeval function.
}
\examples{
f <- stats::runif(20)\%>\%sample(10)\%>\%lazy()

f()
}
