% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripe.R
\name{ripe}
\alias{ripe}
\title{FUNCTION_TITLE}
\usage{
ripe(x, f, ...)
}
\arguments{
\item{x}{chain}

\item{f}{replication function}

\item{\dots}{arguments to pass to f}
}
\value{
object that f outputs
}
\description{
Replicate magrittr chain
}
\examples{
stats::runif(20)\%>\%
 sample(4)\%>\%
 ripe(replicate,n=4,simplify=FALSE)
}
\seealso{
\code{\link[purrr]{rerun}}
}
