% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripe.R
\name{ripe}
\alias{ripe}
\alias{ripely}
\title{Replicate magrittr chain}
\usage{
ripe(x, f, ...)

ripely(x, f, ...)
}
\arguments{
\item{x}{chain}

\item{f}{replication function}

\item{\dots}{arguments to pass to f}
}
\value{
object that f outputs
}
\description{
Lazily replicate magrittr chain
}
\details{
\code{ripely} is for *apply family of functions.
}
\examples{
stats::runif(20)\%>\%
 sample(4)\%>\%
 ripe(replicate,n=4,simplify=FALSE)
 
stats::runif(20)\%>\%
 sample(4)\%>\%
 ripely(lapply,X=1:4)
}
\seealso{
\code{\link[purrr]{rerun}}
}
